requirements.txt
fastapi
python-dotenv
pydantic
langchain_community
langchain
langchain_huggingface